data-pipeline/hpc_etl_pipeline/requirements.txt

pyyaml
watchdog
pandas
pyarrow
beautifulsoup4
requests
